Make usage of inputenc and fontenc optional
When using UTF-8 based engines, such as xelatex from XeTeX, you don't
want to use inputenc and fontenc. With this change, it becomes possible
to pass None for the inputenc and fontenc arguments to Document to avoid
pulling in these packages.

There are no longer any warnings like

*** you should *not* be loading the inputenc package
*** XeTeX expects the source to be in UTF8 encoding
*** some features of other encodings may conflict, resulting in poor output.

when compiling with xelatex.

Fixes #170