Improve whitespace handling in ContainerCommand
ContainerCommand added an extra space character to its output,
and it added newline characters without prefacing them with
a LaTeX percent sign (%) comment character.  These caused extra
whitespace in the output generated by LaTeX.  For many usages
that's okay, but when placing two containers next to one another
in a LaTeX document that can cause extra whitespace in the generated
PDF document, which is not necessarily desirable.  Issue #116 addressed
a similar problem.

This commit squashes the space character and adds two LaTeX comment
characters (%) to eliminate the undesirable whitespace.