Add examples/kitten.jpg to sdist (#379)
The source tarball in PyPI contains tests but `tests/test_pictures.py`
fails because it can't find `examples/kitten.jpg`. The file isn't
contained in the tarball. This PR adds this file to the source tarball
so the tests can be run successfully.

This is relevant when packaging PyLaTeX for distros. For instance, I'm
packaging this for NixOS and the unit tests are run automatically when
distro packages are built. Running the test suite as part of the distro
package building is great and it would be nice if the tests worked
properly.